# web-app/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MomJobs mobile API
  version: 1.0.0
  summary: REST/JSON API for the MomJobs mobile app (candidates and employers).
  description: |
    Plain REST/JSON API that mirrors the MomJobs web app (same business rules, privacy rules, moderation and rate limits).
    Human-readable reference: `docs/api/README.md`, `candidate.md`, `employer.md`, `shared.md`.

    **Headers.** Send `Accept: application/json` (the server also forces it for `/api/*`, so errors are always JSON). Bodies are JSON, except the CV upload
    (`multipart/form-data`). Every endpoint except `auth/register`, `auth/login`, `public/*`, `articles*` and this spec
    requires `Authorization: Bearer <token>` – a Sanctum personal access token from `POST /auth/login` or `POST /auth/register`.
    Tokens expire after 60 days (revoked earlier by logout, logout-all or a password change/reset); one token per device.

    **Rate limits.** Every authenticated endpoint shares a limit of 120 requests per minute per user; some endpoints
    have a stricter limit of their own (noted in their description). A hit returns `429` with a `Retry-After` header.

    **Conventions.** Single resources are wrapped in `data`. Lists are paginated Laravel-style (20 per page,
    `data` / `links` / `meta`, request further pages with `?page=N`); some lists add their own keys to `meta`.
    Small fixed lists are not paginated. Timestamps are ISO 8601, date-only fields `Y-m-d`, workday times `HH:MM`.
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/me/preferences:
    patch:
      tags: [Auth]
      operationId: updatePreferences
      summary: Update account preferences
      description: '`push_enabled: false` stops push notifications on all devices of the account (the devices stay registered).'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [push_enabled]
              properties:
                push_enabled: { type: boolean }
            example: { push_enabled: false }
      responses:
        '200':
          description: The updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /candidate/profile:
    get:
      tags: [Candidate]
      operationId: candidateProfile
      summary: Own full profile
      description: Includes private dates (never shown to employers). The CV file path and raw CV text are never returned.
      responses:
        '200': { $ref: '#/components/responses/CandidateProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /reviews:
    get:
      tags: [Reviews]
      operationId: reviews
      summary: Reviewable companies and your reviews
      description: A candidate may review (once) a company whose invitation she accepted. Not paginated.
      responses:
        '200':
          description: Overview.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [reviewable_companies, reviews]
                    properties:
                      reviewable_companies:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            name: { type: string }
                            city: { type: [string, 'null'] }
                      reviews:
                        type: array
                        items: { $ref: '#/components/schemas/OwnCompanyReview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Laravel Sanctum personal access token (`3|Jt0k…`) from `POST /auth/login` or `POST /auth/register`.
...
  responses:
...
    Unauthorized:
      description: Missing, invalid or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: Unauthenticated. }
    Forbidden:
      description: 'Wrong role, not a participant / owner, unverified e-mail (body adds `email_verification_required: true`, resend the link with `POST /auth/email/verification-notification`) or the action is not allowed in the current state (message may be Polish).'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: This action is unauthorized. }
...
    ValidationFailed:
      description: Validation failed; messages are Polish and can be shown as-is. Moderated texts add a `*_suggestion` key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }
          examples:
            validation:
              value: { message: Pole e-mail jest wymagane., errors: { email: [Pole e-mail jest wymagane.] } }
            moderation:
              value:
                message: Pytania o ciążę i plany rodzinne są niedozwolone.
                errors:
                  body: [Pytania o ciążę i plany rodzinne są niedozwolone.]
                  body_suggestion: ['Od kiedy może Pani rozpocząć pracę i jaki wymiar godzin Pani odpowiada?']
    TooManyRequests:
      description: Rate limit hit.
      headers:
        Retry-After:
          description: Seconds to wait.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { message: Too Many Attempts. }
    CandidateProfile:
      description: The candidate's own profile.
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: '#/components/schemas/CandidateProfile' }
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message: { type: string }
        errors:
          type: object
          description: Field name (dot notation for arrays, e.g. `schedule.0.starts_at`) → list of Polish messages.
          additionalProperties:
            type: array
            items: { type: string }
...
    LabeledValue:
      type: object
      required: [value, label]
      properties:
        value: { type: string }
        label: { type: string, description: 'Polish label' }
    IdName:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...
    DayPart:
      type: string
      enum: [morning, afternoon, any]
...
    User:
      type: object
      required: [id, name, email, email_verified, role, push_enabled, company, candidate_profile]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        role: { type: string, enum: [candidate, employer, admin] }
        push_enabled: { type: boolean, description: 'Push notifications on (default) or off for all devices.' }
        company: { oneOf: [{ $ref: '#/components/schemas/IdName' }, { type: 'null' }] }
        candidate_profile:
          description: Set for candidates only.
          type: [object, 'null']
          properties:
            id: { type: integer }
            published: { type: boolean }
            onboarding_step: { type: integer }
    UserEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/User' }
...
    CandidateProfile:
      type: object
      description: The candidate's own full profile. Contains private fields (`full_name`, `leave_starts_on`, `due_date`, CV info) that are never shown to other users.
      properties:
        id: { type: integer }
        full_name: { type: string }
        anonymous_name: { type: string, example: Marta K. }
        headline: { type: [string, 'null'] }
        years_of_experience: { type: [integer, 'null'] }
        city: { type: [string, 'null'] }
        phone: { type: [string, 'null'], example: '+48 600 100 200', description: 'Private – revealed only to a company whose invitation she accepted.' }
        photo_url: { type: [string, 'null'], format: uri, description: 'Authorised endpoint `GET /candidate-photos/{profile}` (send the bearer token); `null` without a photo.' }
        ai_summary: { type: [string, 'null'] }
        available_from: { type: [string, 'null'], format: date }
        leave_starts_on: { type: [string, 'null'], format: date, description: 'Private.' }
        due_date: { type: [string, 'null'], format: date, description: 'Private.' }
        work_modes: { type: array, items: { $ref: '#/components/schemas/LabeledValue' } }
        employment_fractions: { type: array, items: { $ref: '#/components/schemas/LabeledValue' } }
        wants_flexible_hours: { type: boolean }
        open_to_job_sharing: { type: boolean }
        preferred_day_part: { oneOf: [{ $ref: '#/components/schemas/DayPart' }, { type: 'null' }] }
        preferred_day_part_label: { type: [string, 'null'] }
        privacy:
          type: object
          properties:
            show_availability_instead_of_gap: { type: boolean }
            allow_direct_messages: { type: boolean }
            job_alerts_enabled: { type: boolean, description: 'Weekly e-mail with new offers matching at least 60%.' }
            hidden_from_company: { oneOf: [{ $ref: '#/components/schemas/IdName' }, { type: 'null' }] }
        cv:
          type: object
          description: The CV file path and raw text are never returned.
          properties:
            original_name: { type: [string, 'null'] }
            size: { type: [integer, 'null'], description: 'Bytes.' }
            status: { type: [string, 'null'], enum: [uploaded, parsing, parsed, failed, null] }
            has_text: { type: boolean }
        skills:
          type: array
          description: Only `confirmed` skills are visible to employers and used for matching.
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              source: { type: [string, 'null'], enum: [ai, manual, null] }
              confirmed: { type: boolean }
        suggested_positions:
          type: array
          items:
            type: object
            properties:
              title: { type: string }
              score: { type: integer }
        onboarding_step: { type: integer }
        published: { type: boolean }
        published_at: { type: [string, 'null'], format: date-time }
...
    OwnCompanyReview:
      type: object
      description: The candidate's own review with its moderation status.
      properties:
        id: { type: integer }
        company: { $ref: '#/components/schemas/IdName' }
        rating_return: { type: integer, minimum: 1, maximum: 5 }
        rating_flexibility: { type: integer, minimum: 1, maximum: 5 }
        rating_no_pregnancy_questions: { type: integer, minimum: 1, maximum: 5 }
        quote: { type: string }
        author_label: { type: [string, 'null'] }
        status: { type: string, enum: [pending, approved, rejected] }
        status_label: { type: string }
        can_edit: { type: boolean }
        created_at: { type: [string, 'null'], format: date-time }
...